SSMW-215 scripts: upload package to test in Nexus
Anonymous right access has been removed in bamboo. Hence, to
be able to run the LAVA CI, the package file has to be uploaded
on Nexus to allow the linux running on board to get it.

Add a CI task to clean up Nexus folder at the end of the PR/branch
verification.

Signed-off-by: Cedric Neveux <cedric.neveux@nxp.com>