SSMW-242 tests: Fix test definition version parameter
Some test definition file use a parameter "version" but it is set as a
string and not an integer.

Change the test definition file U_API_Info_001 to use a "lib_version"
parameter instead of "version" parameter. This is not the same type
of data.
The library version is handled via a double data type.

Signed-off-by: Cedric Neveux <cedric.neveux@nxp.com>