SSMW-392 core: keymgr: Implement get key lengths implicit operation
Implementation of the implicit operation get key lengths.
This implicit operation is implemented in the subsystem howning the key
given in parameter of the smw_get_key_buffers_length().

Add functions smw_keymgr_update_XXX_buffer() to update the user buffers
data and length fields depending of the buffer format base64 or not.

Signed-off-by: Cedric Neveux <cedric.neveux@nxp.com>