SSMW-492 scripts: Allow to install TEE TAs in dedicated directory
Add a build option TEE_TA_DESTDIR allowing to install TEE TA in
a dedicated directory where the optee_armtz is created.

If TEE_TA_DESTDIR is not defined the default path where is installed the
TA is `[DESTDIR]/lib/optee_armtz`

if TEE_TA_DESTDIR is defined the TA is installed in the
the path `[DESTDIR]/[TEE_TA_DESTDIR]/optee_armtz`

The variable `DESTDIR` is the cmake environment variable that could be
specified during the install cmake command.
E.g. `make install DESTDIR=tmp`

Signed-off-by: Cedric Neveux <cedric.neveux@nxp.com>