SSMW-112 pkcs11: Add support to message based encryption decryption
Implement
- C_MessageEncryptInit()
- C_EncryptMessage()
- C_EncryptMessageBegini()
- C_EncryptMessageNext()
- C_MessageEncryptFinal()
- C_MessageDecryptInit()
- C_DecryptMessage()
- C_DecryptMessageBegin()
- C_DecryptMessageNext()
- C_MessageDecryptFinal()

Modification
- Copy libcipher iv and tag buffer from user parameter:
The IV in the struct lib_cipher_ctx context is a reference to the user
application parameter and it's set when C_[Encrypt/Decrypt]Init are called.
Problem is that as it's user parameter, we can't guaranty that the
IV buffer will remain valid until we effectively use it.
We need to allocate the IV of the struct lib_cipher_ctx context and copy
the user IV. Same scenario for user application Tag buffer.

Signed-off-by: Olivier Masse <olivier.masse@nxp.com>