SSMW-713 core: storage: Find free data id
User can request storing a data with a data unique id set to 0,
it's the case for PKCS11.
As subsystem expects to have a data id already set before storing it,
SMW creates a new database object to get a free object id, then query
each subsystem to check if the data id already exists or not.
If yes, increments the data id.

Signed-off-by: Cedric Neveux <cedric.neveux@nxp.com>