SSMW-719 core: check key type/algo on generate
There may be multiple subsystems which support an operation, but not
necessarily with the same key types/modes. Consider this example:

[SECURE_SUBSYSTEM]
    TEE;
    [SECURITY_OPERATION]
        GENERATE_KEY;
    [SECURITY_OPERATION]
        AEAD;
        KEY_TYPE_VALUES=AES;
        MODE_VALUES=CCM;
[SECURE_SUBSYSTEM]
    ELE;
    [SECURITY_OPERATION]
        GENERATE_KEY;
    [SECURITY_OPERATION]
    AEAD;
        KEY_TYPE_VALUES=AES;
        MODE_VALUES=CCM:GCM:CHACHA20_POLY1305;

Currently, when a key with type=AES and mode=POLY1305 is being
generated, it would be assigned to TEE, even though TEE does not support
it. GENERATE_KEY can only restrict based on key type, not key mode/algo.

So, during key generation, also search through a subsystem's operations
to see if there is one that uses the key mode/algo explicitly.

Signed-off-by: Ilie Halip <ilie.halip@nxp.com>