SSMW-813 subsystem: tee: Improve objects storage management
ta_find_unused_object_id was calling ta_find_and_open_persistent_id for all
possible ids from 0 to 2^32.
This last one was doing enumerations of all persistent objects in the
secure filesystem.

As a result, for N created objects in the TA, we do N^2 enumeration in
OPTEE secure filesystem.

O(n^2) complexity caused performance issue.
This is solved by keeping a list of persistent objects and get the first
free id in this list.

Signed-off-by: Olivier Masse <olivier.masse@nxp.com>