SSMW-789 subsystems: tee: Set the correct hash buffer length
If the hash buffer is NULL and hash buffer length is not 0,
TEE_DigestDoFinal() returns `TEE_ERROR_BAD_PARAMETERS`.
This change ensures that when the hash buffer is NULL,
the length is explicitly set to 0, allowing the function to
return `TEE_ERROR_SHORT_BUFFER` and report the required digest size.

Signed-off-by: Rakshata Nayak <rakshata.nayak@nxp.com>